Move Home and Navbar components to components folder
