+++ src/routes/Routes.js
@@ -3,9 +3,9 @@ import React from 'react';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import Categories from '../containers/Categories';
 import Countries from '../containers/Countries';
-import Home from '../containers/Home';
+import Home from '../components/Home';
 import MealList from '../containers/MealList';
-import Navbar from '../containers/Navbar';
+import Navbar from '../components/Navbar';
 import Recipe from '../containers/Recipe';
 
 const Routes = () => (
